[VL] Fix passing Velox session configurations

+++ cpp/velox/compute/WholeStageResultIterator.cc
@@ -618,7 +618,7 @@ std::unordered_map<std::string, std::string> WholeStageResultIterator::getQueryC
     configs[velox::core::QueryConfig::kAbandonBuildNoDupHashMinRows] =
         std::to_string(veloxCfg_->get<int32_t>(kAbandonBuildNoDupHashMinRows, 100000));
     configs[velox::core::QueryConfig::kAbandonBuildNoDupHashMinPct] =
-        std::to_string(veloxCfg_->get<int32_t>(kAbandonBuildNoDupHashMinPct, 0));
+        std::to_string(static_cast<int32_t>(veloxCfg_->get<double>(kAbandonBuildNoDupHashMinPct, 0)));
 
     // Disable driver cpu time slicing.
     configs[velox::core::QueryConfig::kDriverCpuTimeSliceLimitMs] = "0";
